fix(techArticle): 기술블로그 키워드검색시 score 응답 누락 수정

+++ src/main/java/com/dreamypatisiel/devdevdev/domain/service/techArticle/techArticle/TechArticleService.java
@@ -11,7 +11,7 @@ import org.springframework.security.core.Authentication;
 
 public interface TechArticleService {
     Slice<TechArticleMainResponse> getTechArticles(Pageable pageable, Long techArticleId, TechArticleSort techArticleSort,
-                                                   String keyword, Long companyId, Float score,
+                                                   String keyword, Long companyId, Double score,
                                                    Authentication authentication);
 
     TechArticleDetailResponse getTechArticle(Long techArticleId, String anonymousMemberId, Authentication authentication);
