+++ src/test/java/com/dreamypatisiel/devdevdev/web/docs/NotificationControllerDocsTest.java
@@ -510,7 +510,7 @@ class NotificationControllerDocsTest extends SupportControllerDocsTest {
                                                                 String techArticleUrl, String title, String contents,
                                                                 Long companyId, String companyName, String careerUrl, String officialImageUrl,
                                                                 LocalDate regDate, String author, long recommendCount,
-                                                                long commentCount, long viewCount, Boolean isBookmarked, Float score) {
+                                                                long commentCount, long viewCount, Boolean isBookmarked, Double score) {
       return TechArticleMainResponse.builder()
               .id(id)
               .thumbnailUrl(thumbnailUrl)
